@@ -13,50 +13,52 @@
   "repository": "https://github.com/blankoslo/floq-balance",
   "license": "UNLICENSED",
   "devDependencies": {
-    "babel-core": "^6.2.1",
-    "babel-eslint": "^6.0.2",
-    "babel-loader": "^6.2.0",
+    "babel-core": "^6.14.0",
+    "babel-eslint": "^6.1.2",
+    "babel-loader": "^6.2.5",
     "babel-preset-airbnb": "^2.0.0",
-    "babel-preset-es2015": "^6.1.18",
-    "babel-preset-react": "^6.1.18",
-    "babel-preset-stage-1": "^6.1.18",
-    "css-loader": "^0.23.1",
-    "enzyme": "^2.0.0",
-    "eslint": "^2.7.0",
-    "eslint-config-airbnb": "^9.0.0",
-    "eslint-loader": "^1.3.0",
-    "eslint-plugin-import": "^1.5.0",
-    "eslint-plugin-jsx-a11y": "^1.0.2",
-    "eslint-plugin-react": "^5.0.1",
-    "expect": "^1.16.0",
-    "fetch-mock": "^4.6.0",
-    "jsdom": "^9.0.0",
-    "less": "^2.6.1",
-    "less-loader": "^2.2.2",
-    "mocha": "^2.4.5",
-    "react-addons-perf": "^15.2.1",
-    "react-addons-test-utils": "^15.0.0",
+    "babel-preset-es2015": "^6.14.0",
+    "babel-preset-react": "^6.11.1",
+    "babel-preset-stage-1": "^6.13.0",
+    "css-loader": "^0.24.0",
+    "enzyme": "^2.4.1",
+    "eslint": "^3.4.0",
+    "eslint-config-airbnb": "^10.0.1",
+    "eslint-import-resolver-webpack": "^0.5.1",
+    "eslint-loader": "^1.5.0",
+    "eslint-plugin-import": "^1.14.0",
+    "eslint-plugin-jsx-a11y": "^2.2.0",
+    "eslint-plugin-react": "^6.2.0",
+    "expect": "^1.20.2",
+    "fetch-mock": "^5.1.1",
+    "jsdom": "^9.4.2",
+    "less": "^2.7.1",
+    "less-loader": "^2.2.3",
+    "mocha": "^3.0.2",
+    "react-addons-perf": "^15.3.1",
+    "react-addons-test-utils": "^15.3.1",
     "react-hot-loader": "^1.3.0",
-    "redux-devtools": "^3.1.1",
-    "redux-devtools-dock-monitor": "^1.1.0",
-    "redux-devtools-log-monitor": "^1.0.5",
-    "redux-mock-store": "^1.0.3",
-    "style-loader": "^0.13.0",
-    "webpack": "^1.12.14",
-    "webpack-dev-server": "^1.14.0"
+    "redux-devtools": "^3.3.1",
+    "redux-devtools-dock-monitor": "^1.1.1",
+    "redux-devtools-log-monitor": "^1.0.11",
+    "redux-mock-store": "^1.1.4",
+    "style-loader": "^0.13.1",
+    "webpack": "^1.13.2",
+    "webpack-dev-server": "^1.15.0"
   },
   "dependencies": {
     "immutable": "^3.8.1",
-    "material-ui": "^0.15.2",
+    "material-ui": "^0.15.4",
     "moment": "^2.14.1",
-    "react": "^15.2.1",
-    "react-addons-pure-render-mixin": "^15.2.1",
-    "react-dom": "^15.2.1",
-    "react-redux": "^4.0.0",
-    "react-router": "^2.0.0",
+    "react": "^15.3.1",
+    "react-addons-pure-render-mixin": "^15.3.1",
+    "react-dom": "^15.3.1",
+    "react-redux": "^4.4.5",
+    "react-router": "^2.7.0",
+    "react-router-redux": "^4.0.5",
     "react-tap-event-plugin": "^1.0.0",
-    "redux": "^3.0.4",
-    "reselect": "^2.5.1",
+    "redux": "^3.5.2",
+    "reselect": "^2.5.3",
     "whatwg-fetch": "^1.0.0"
   }
 }
