@@ -1,13 +1,13 @@
 import React from 'react';
 
-const BalanceViewBodyCell = (props) => (
+const BalanceViewBodyCell = props => (
   <td>
     {props.value}
   </td>
 );
 
 BalanceViewBodyCell.propTypes = {
-  value: React.PropTypes.string
+  value: React.PropTypes.string.isRequired
 };
 
 export default BalanceViewBodyCell;
