@@ -1,58 +1,51 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import { getProjects } from '../actions/index';
+import { getProjects, getHoursPerProject } from '../actions/index';
+import titleSelector from '../selectors/titleSelector';
+import tableBodySelector from '../selectors/tableBodySelector';
+import IndexComponent from '../components/index';
 
 class App extends Component {
   constructor(props) {
     super(props);
 
     props.getProjects();
+    props.getHoursPerProject(props.title.startDate, props.title.endDate);
   }
 
-  // componentWillReceiveProps(nextProps) {
-  //   if (nextProps.location.query.year !== this.props.location.query.year) {
-  //     this.props.selectYear(nextProps.location.query.year);
-  //   }
-  //   if (nextProps.location.query.week !== this.props.location.query.week) {
-  //     this.props.selectWeek(nextProps.location.query.week);
-  //   }
-  // }
+  componentWillReceiveProps(nextProps) {
+    if (nextProps.title.startDate !== this.props.title.startDate ||
+        nextProps.title.endDate !== this.props.title.endDate) {
+      nextProps.getHoursPerProject(nextProps.title.startDate, nextProps.title.endDate);
+    }
+  }
 
   render() {
-    if (this.props.projects.loading) {
+    if (this.props.tableBody.loading) {
       return null;
     }
-    const children = React.Children.map(this.props.children,
-      child => React.cloneElement(child, {
-        projects: this.props.projects,
-      }));
-
-    return (
-      <div>
-        {children}
-      </div>
-    );
+    return <IndexComponent tableBody={this.props.tableBody.data} title={this.props.title} />;
   }
 }
 
 App.propTypes = {
-  location: React.PropTypes.object.isRequired,
-  params: React.PropTypes.object.isRequired,
-  children: React.PropTypes.object,
-
   // mapStateToProps
-  projects: React.PropTypes.object.isRequired,
+  tableBody: React.PropTypes.object.isRequired,
+  title: React.PropTypes.object.isRequired,
 
   // mapDispatchToProps
   getProjects: React.PropTypes.func.isRequired,
+  getHoursPerProject: React.PropTypes.func.isRequired,
 };
 
-const mapStateToProps = (state) => ({
-  projects: state.projects,
+const mapStateToProps = (state, ownProps) => ({
+  tableBody: tableBodySelector(state, ownProps),
+  title: titleSelector(state, ownProps),
 });
 
 const mapDispatchToProps = {
   getProjects,
+  getHoursPerProject
 };
 
 export default connect(mapStateToProps, mapDispatchToProps)(App);
