@@ -1,14 +1,14 @@
 import * as Immutable from 'immutable';
 
-import { GET_PROJECTS } from '../actions/index';
+import { GET_HOURS_PER_PROJECT } from '../actions/index';
 
 export default (state = { loading: true, data: new Immutable.Map() }, action) => {
   switch (action.type) {
-    case GET_PROJECTS:
+    case GET_HOURS_PER_PROJECT:
       return {
         loading: false,
-        data: new Immutable.OrderedMap(action.payload.map(p => [p.id, p]))
-            .sortBy(p => p.id.toLowerCase())
+        data: new Immutable.OrderedMap(action.payload.map(e => [e.project, e]))
+            .sortBy(e => e.project.toLowerCase())
       };
     default:
       return state;
